add tests for running experiments script
